Revert "Port Bank Daemon Pt1"
This reverts commit 2da674afd3646375edf8607c4db781709150dce1.
